create weapon info service just to make sure that everything isn't breaking horribly

+++ WycademyV2/src/WycademyV2/Commands/Utilities/IsDeviantWeaponConverter.cs
@@ -18,7 +18,7 @@ namespace WycademyV2.Commands.Utilities
 
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
-            var input = (int)reader.Value;
+            var input = (long)reader.Value;
 
             return input == 1;
         }
